改动：变频器参数正常上传，但上传完成后变频器报警 日期：2017.12.05

+++ Inc/cp.h
@@ -194,7 +194,7 @@ typedef struct _cp
     u16 disp_para2; //主界面显示参数2
     u16 disp_para3; //主界面显示参数3
     u16 vfd_para; //变频器参数
-    fp32 vfd_para_count; //变频器参数当前数目
+    u16 vfd_para_count; //变频器参数当前数目
     u16 vfd_para_total; //变频器参数总数目
 } CP, *PCP;
